+++ examples/embedded-wallet-service-sdk-demo-app/CHANGELOG.md
@@ -1,5 +1,16 @@
 # embedded-wallet-service-sdk-demo-app
 
+## 0.1.2
+
+### Patch Changes
+
+- 7440c1b: feat(embedded-wallet-service-sdk): Add ability to support any arbitrary evm chain via 'rpcEndpoint`override when calling`userwallet.getEthersJsSigner({ rpcEndpoint: "" })`
+
+  chore(sdk-common-utilities): update public rpc endpoints to more generous ones
+
+- Updated dependencies [7440c1b]
+  - @paperxyz/embedded-wallet-service-sdk@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes