+++ packages/react-client-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @paperxyz/react-client-sdk
 
+## 0.9.2
+
+### Patch Changes
+
+- 1caa827: Allow setting the RPC URL(s) for CheckoutWithEth
+
+## 0.9.1
+
+### Patch Changes
+
+- 542ac25: Pin JS SDK dependencies
+- Updated dependencies [7440c1b]
+  - @paperxyz/sdk-common-utilities@0.0.3
+
 ## 0.9.0
 
 ### Minor Changes